agent/requirements.txt
anthropic
openai
google-generativeai
browser-use
agentmail
supermemory
laminar
paylocus
httpx
